backend/requirements.txt
vision-agents
opencv-python
numpy
websockets
asyncio
python-dotenv
ultralytics
Pillow
fastapi
uvicorn
